refactor: move addTable to subfolder table
