+++ app/services/pdf/fluggastrechte/sections/secondPage/factsOfCases/createFactsOfCases.ts
@@ -3,7 +3,7 @@ import { addDetailedReason } from "./addDetailedReason";
 import { addFlightDetails } from "./addFlightDetails";
 import { addLastSentences } from "./addLastSentences";
 import { addReason } from "./addReason";
-import { addTable } from "./addTable";
+import { addTable } from "./table/addTable";
 import { FONTS_BUNDESSANS_BOLD } from "../../../createPdfKitDocument";
 
 export const createFactsOfCases = (
